refactor: isolate code in specific class Signed-off-by: Philippe Martin <phmartin@redhat.com>
